servicio_pedidos/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0
SQLAlchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aio-pika>=9.0.0
alembic
PyJWT